completed upload and serving files features

+++ controllers/admin.js
@@ -5,9 +5,9 @@ const { validationResult } = require('express-validator/check');
 const Product = require('../models/product');
 
 exports.getAddProduct = (req, res, next) => {
-    if (!req.session.isLoggedIn) {
-        return res.redirect('/login');
-    }
+    // if (!req.session.isLoggedIn) {
+    //     return res.redirect('/login');
+    // }
     res.render('admin/edit-product', 
         {
             pageTitle: 'Add Product', 
@@ -24,29 +24,29 @@ exports.postAddProduct = (req, res, next) => {
     const image = req.file;
     const price = req.body.price;
     const description = req.body.description;
-    const imageUrl = image.path;
+    console.log("Post Add Product", image);
     if (!image) {
         return res.status(422).render('admin/edit-product', {
-          pageTitle: 'Add Product',
-          path: '/admin/add-product',
-          editing: false,
-          hasError: true,
-          product: {
-            title: title,
-            price: price,
-            description: description
-          },
-          errorMessage: 'Attached file is not an image.',
-          validationErrors: []
+            pageTitle: 'Add Product',
+            path: '/admin/add-product',
+            editing: false,
+            hasError: true,
+            product: {
+                title: title,
+                price: price,
+                description: description
+            },
+            errorMessage: 'Attached file is not an image.',
+            validationErrors: []
         });
-      }
-      const errors = validationResult(req);
+    }
+
+    const errors = validationResult(req);
 
     if (!errors.isEmpty()) {
         console.log(errors.array());
-        return res.status(422).render('admin/edit-product', 
-        {
-            pageTitle: 'Add Product', 
+        return res.status(422).render('admin/edit-product', {
+            pageTitle: 'Add Product',
             path: '/admin/add-product',
             editing: false,
             hasError: true,
@@ -61,10 +61,10 @@ exports.postAddProduct = (req, res, next) => {
         });
     }
 
-    
+    const imageUrl = image.path;
 
     const product = new Product({
-        //_id: new mongoose.Types.ObjectId('5db0787f7a7488141c0ad4fa'),
+        // _id: new mongoose.Types.ObjectId('5badf72403fd8b5be0366e81'),
         title: title,
         price: price,
         description: description,
@@ -72,34 +72,32 @@ exports.postAddProduct = (req, res, next) => {
         userId: req.user
     });
     product
-    .save()
-    .then(result => {
-        //console.log(result);
-        console.log('Created Product');
-        res.redirect('/admin/products');
-    })
-    .catch(err => {
-        // return res.status(500).render('admin/edit-product', 
-        // {
-        //     pageTitle: 'Add Product', 
-        //     path: '/admin/add-product',
-        //     editing: false,
-        //     hasError: true,
-        //     product: {
-        //         title: title,
-        //         imageUrl: imageUrl,
-        //         price: price,
-        //         description: description
-        //     },
-        //     errorMessage: 'Database operation failed, please try again.',
-        //     validationErrors: []
-        // });
-       
-        //res.redirect('/500');    
-        const error = new Error(err);
-        error.httpStatusCode = 500;        
-        return next(error);
-    });
+        .save()
+        .then(result => {
+            // console.log(result);
+            console.log('Created Product');
+            res.redirect('/admin/products');
+        })
+        .catch(err => {
+            // return res.status(500).render('admin/edit-product', {
+            //   pageTitle: 'Add Product',
+            //   path: '/admin/add-product',
+            //   editing: false,
+            //   hasError: true,
+            //   product: {
+            //     title: title,
+            //     imageUrl: imageUrl,
+            //     price: price,
+            //     description: description
+            //   },
+            //   errorMessage: 'Database operation failed, please try again.',
+            //   validationErrors: []
+            // });
+            // res.redirect('/500');
+            const error = new Error(err);
+            error.httpStatusCode = 500;
+            return next(error);
+        });
 };
 
 exports.getEditProduct = (req, res, next) => {
@@ -143,17 +141,15 @@ exports.postEditProduct = (req, res, next) => {
 
     if (!errors.isEmpty()) {
         console.log(errors.array());
-        return res.status(422).render('admin/edit-product', 
-        {
-            pageTitle: 'Edit Product', 
-            path: '/admin/edit-product',
-            editing: true,
+        return res.status(422).render('admin/edit-product', {
+            pageTitle: 'Add Product',
+            path: '/admin/add-product',
+            editing: false,
             hasError: true,
             product: {
-                title: updatedTitle,
-                price: updatedPrice,
-                description: updatedDesc,
-                _id: prodId
+                title: title,
+                price: price,
+                description: description
             },
             errorMessage: errors.array()[0].msg,
             validationErrors: errors.array()
@@ -169,14 +165,14 @@ exports.postEditProduct = (req, res, next) => {
             product.price = updatedPrice;
             product.description = updatedDesc;
             if (image) {
+                fileHelper.deleteFile(product.imageUrl);
                 product.imageUrl = image.path;
-              }
-            return product.save()
-                .then(result => {
+            }
+            return product.save().then(result => {
                 console.log('UPDATED PRODUCT!');
                 res.redirect('/admin/products');
             });
-        })        
+        })
         .catch(err => {
             const error = new Error(err);
             error.httpStatusCode = 500;
@@ -205,14 +201,23 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
     const prodId = req.body.productId;
-    Product.deleteOne({_id: prodId, userId: req.user._id })
-    .then(() => {
+    Product.findById(prodId)
+      .then(product => {
+        if (!product) {
+          return next(new Error('Product not found.'));
+        }
+        if(product.imageUrl) {
+          fileHelper.deleteFile(product.imageUrl);
+        }
+        return Product.deleteOne({ _id: prodId, userId: req.user._id });
+      })
+      .then(() => {
         console.log('DESTROYED PRODUCT');
         res.redirect('/admin/products');
-    })
-    .catch(err => {
+      })
+      .catch(err => {
         const error = new Error(err);
         error.httpStatusCode = 500;
         return next(error);
-    });
-};
+      });
+  };
